Integrate scale sub resouce by means of an HPA and ScheludedScaling CRD
This pr development focusses on using the a scaling schedule or a
cluster scaling schedule CRD's for automatically scaling up an EDS to
the required number of replicas. The HPA doesn't directly control or
edit the replicas of the EDS. The replic count reported by the HPA is
stored in a separate property which is taken into account during the
calculation for determining the scaling operation.

Signed-off-by: Girish Chandrashekar <girish.chandrashekar@zalando.de>